% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{bqr}
\alias{bqr}
\title{Cleaned version of the "bqr" dataset}
\format{
A data frame with 358 observations of the protein sequence "A".
}
\usage{
bqr
}
\description{
This polymerase dataset contains the cleaned version of the "bqr" dataset.
}
\examples{
data(bqr)
head(bqr)
}
\seealso{
\code{\link{bqr}}
}
\keyword{datasets}
