% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classifies Proteins}
\usage{
classify(data)
}
\arguments{
\item{data}{Input protein string of amino acids}
}
\value{
returns data frame with amino acid and its classification
}
\description{
Takes a vector and will characterize each amino acid with a charge or
polarity
}
