% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{cleaner}
\alias{cleaner}
\title{Cleans pdb file of proteins}
\usage{
cleaner(pdb_location)
}
\arguments{
\item{pdb_location}{Input location of pdb file on local device}
}
\value{
returns string that is longer of A or B protein sequence
}
\description{
Cleans pdb file of proteins
}
