% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{fpu}
\alias{fpu}
\title{Cleaned version of the "fpu" dataset}
\format{
A data frame with 293 observations of the protein sequence "A".
}
\usage{
fpu
}
\description{
This kinase dataset contains the cleaned version of the "fpu" dataset.
}
\examples{
data(fpu)
head(fpu)
}
\seealso{
\code{\link{fpu}}
}
\keyword{datasets}
