% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{huo}
\alias{huo}
\title{Cleaned version of the "huo" dataset}
\format{
A data frame with 335 observations of the protein sequence "A".
}
\usage{
huo
}
\description{
This polymerase dataset contains the cleaned version of the "huo" dataset.
}
\examples{
data(huo)
head(huo)
}
\seealso{
\code{\link{huo}}
}
\keyword{datasets}
