% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{huo1}
\alias{huo1}
\title{Raw version of the "huo" dataset}
\format{
A data frame the entire protein structure.
}
\usage{
huo1
}
\description{
This polymerase dataset contains the raw version of the "huo" dataset.
}
\seealso{
\code{\link{huo1}}
}
\keyword{datasets}
