% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.protein5555}
\alias{plot.protein5555}
\title{Plot protein information}
\usage{
\method{plot}{protein5555}(x, y, kind = "aminoDis", ...)
}
\arguments{
\item{x}{protein5555 object}

\item{y}{another parameter}

\item{kind}{"aminoDis", "charge", "polar"}
}
\description{
Ability to plot
}
