% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein5555.R
\name{protein5555}
\alias{protein5555}
\title{Create a protein constructor}
\usage{
protein5555(sequence, family)
}
\arguments{
\item{sequence}{Requires a protein sequence to put inputted}

\item{family}{Requires the family, either polymerase or kinase, to be input}
}
\description{
Create a protein constructor
}
