% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein5555.R
\name{protein5555_list}
\alias{protein5555_list}
\title{Create a protein list constructor}
\usage{
protein5555_list(proteins)
}
\arguments{
\item{proteins}{Input is a list of proteins to create an overall list of
proteins}
}
\description{
Create a protein list constructor
}
