% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{qgy}
\alias{qgy}
\title{Cleaned version of the "qgy" dataset}
\format{
A data frame with 896 observations of the protein sequence "A".
}
\usage{
qgy
}
\description{
This kinase dataset contains the cleaned version of the "qgy" dataset.
}
\examples{
data(gqy)
head(qgy)
}
\seealso{
\code{\link{qgy}}
}
\keyword{datasets}
