% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{r7g}
\alias{r7g}
\title{Cleaned version of the "r7g" dataset}
\format{
A data frame with 301 observations of the protein sequence "A".
}
\usage{
r7g
}
\description{
This kinase dataset contains the cleaned version of the "r7g" dataset.
}
\examples{
data(r7g)
head(r7g)
}
\seealso{
\code{\link{r7g}}
}
\keyword{datasets}
