% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{sl0}
\alias{sl0}
\title{Cleaned version of the "sl0" dataset}
\format{
A data frame with 698 observations of the protein sequence "A".
}
\usage{
sl0
}
\description{
This polymerase dataset contains the cleaned version of the "sl0" dataset.
}
\examples{
data(sl0)
head(sl0)
}
\seealso{
\code{\link{sl0}}
}
\keyword{datasets}
