% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classifies String of Protein into Families}
\usage{
classify(data)
}
\arguments{
\item{data}{Input cleansed protein sequence}
}
\value{
protein data frame of labels of the family type
}
\description{
Classifies String of Protein into Families
}
